fix: receive test
