+++ contracts/cosmwasm-vm/cw-ibc-core/src/ics04_channel/handler/validate_channel.rs
@@ -1,6 +1,6 @@
 use common::ibc::core::ics04_channel::channel::{ChannelEnd, State};
 use common::ibc::core::ics24_host::identifier::ChannelId;
-use cw_common::ibc_types::{ChannelError};
+use cw_common::ibc_types::ChannelError;
 
 use crate::ContractError;
